@@ -3,10 +3,10 @@
     "description": "Validation resources and rules.",
     "type": "project",
     "require": {
-        "php": ">=7.4.2",
+        "php": ">=8.1",
         "ext-json": "*",
         "ext-mbstring": "*",
-        "jitesoft/exceptions": "2.*"
+        "jitesoft/exceptions": "^2.4.2"
     },
     "license": "mit",
     "authors": [
@@ -26,11 +26,17 @@
         }
     },
     "require-dev": {
-        "phpunit/phpunit": "^9.4",
-        "fzaninotto/faker": "^v1.9.1",
-        "squizlabs/php_codesniffer": "^3.5"
+        "phpunit/phpunit": "^9.5",
+        "fakerphp/faker": "^v1.20",
+        "squizlabs/php_codesniffer": "^3.7"
     },
     "scripts": {
+        "lint": [
+            "phpcs --report=full --standard=./ruleset.xml -p ./src"
+        ],
+        "lint:fix": [
+            "phpcbf --standard=./ruleset.xml -p ./src"
+        ],
         "test": [
             "phpunit --configuration ./phpunit.xml --coverage-text --colors=never",
             "phpcs --report=summary -n --standard=./ruleset.xml -p ./src"
