@@ -16,21 +16,19 @@ use Jitesoft\Validator\Tests\AbstractTestCase;
  * @version 1.0.0
  */
 class PatternTest extends AbstractTestCase {
-
-    /** @var RuleInterface */
-    protected $rule;
+    protected RuleInterface $rule;
 
     protected function setUp(): void {
         parent::setUp();
         $this->rule = (new Factory())->create(Pattern::class);
     }
 
-    public function testGetError() {
+    public function testGetError(): void {
         $this->rule->test('abc', ['pattern' => '/abc(\d{3})/i']);
         $this->assertEquals(['pattern' => 'Value did not match supplied pattern.'], $this->rule->popErrors());
     }
 
-    public function testTest() {
+    public function testTest(): void {
         $this->assertTrue($this->rule->test('abc123', ['pattern' => '/abc(\d{3})/i']));
         $this->assertFalse($this->rule->test('abcd123', ['pattern' => '/abc(\d{3})/i']));
     }
