@@ -34,7 +34,7 @@ class Length extends AbstractRule {
      * @return boolean
      * @since 1.0.0
      */
-    protected function testRule($value,
+    protected function testRule(mixed $value,
                                 array $rules = [],
                                 array $args = []): bool {
         $result = true;
