@@ -29,7 +29,8 @@ class Minimum extends AbstractRule {
      * @return boolean
      * @since 1.0.0
      */
-    protected function testRule($value,
+    // phpcs:ignore Generic.CodeAnalysis.UnusedFunctionParameter
+    protected function testRule(mixed $value,
                                 array $rules = [],
                                 array $args = []): bool {
         if (!is_numeric($value)) {
